Develop

+++ Backend/Greenhouse.Application/Mqtt/Dtos/DeviceLogDto.cs
@@ -5,7 +5,7 @@ namespace Greenhouse.Application.Mqtt.Dtos;
 public class DeviceLogDto
 {
     [Required] [MinLength(1)] public string Unit { get; set; }
-    [Required] [MinLength(1)] public int Value { get; set; }
+    [Required] [MinLength(1)] public decimal Value { get; set; }
     [Required] [MinLength(1)] public Guid DeviceId { get; set; }
     [Required] [MinLength(1)] public string Type { get; set; }
 }
